+++ packages/jsActions/mobile-resources-native/src/camera/TakePicture.ts
@@ -56,7 +56,7 @@ export async function TakePicture(
     }
 
     // V3 dropped the feature of providing an action sheet so users can decide on which action to take, camera or library.
-    const nativeVersionMajor = NativeModules.ImagePickerManager.showImagePicker ? 2 : 4;
+    const nativeVersionMajor = NativeModules?.ImagePickerManager?.showImagePicker ? 2 : 4;
     const RNPermissions = nativeVersionMajor === 4 ? (await import("react-native-permissions")).default : null;
 
     try {