+++ packages/pluggableWidgets/bottom-sheet-native/src/components/ExpandingDrawer.tsx
@@ -1,7 +1,7 @@
-import { BottomSheetStyle } from "../ui/Styles";
-import { createElement, ReactNode, useCallback, useState, ReactElement, Children } from "react";
-import BottomSheet from "reanimated-bottom-sheet";
+import { createElement, ReactNode, ReactElement, useCallback, useState, useRef, Children } from "react";
 import { Dimensions, LayoutChangeEvent, SafeAreaView, StyleSheet, View } from "react-native";
+import BottomSheet, { BottomSheetView } from "@gorhom/bottom-sheet";
+import { BottomSheetStyle } from "../ui/Styles";
 
 interface ExpandingDrawerProps {
     smallContent?: ReactNode;
@@ -11,22 +11,25 @@ interface ExpandingDrawerProps {
     onClose?: () => void;
     styles: BottomSheetStyle;
 }
+let lastIndexRef = -1;
+
+const OFFSET_BOTTOM_SHEET = 25;
 
 export const ExpandingDrawer = (props: ExpandingDrawerProps): ReactElement => {
     const [heightContent, setHeightContent] = useState(0);
     const [heightHeader, setHeightHeader] = useState(0);
     const [fullscreenHeight, setFullscreenHeight] = useState(0);
-    const [isOpen, setIsOpen] = useState<boolean>(false);
+    const [isOpen, setIsOpen] = useState<boolean>(true);
+    const bottomSheetRef = useRef<BottomSheet>(null);
+
     const maxHeight = Dimensions.get("screen").height;
     const isSmallContentValid = Children.count(props.smallContent) > 0;
     const isLargeContentValid = Children.count(props.largeContent) > 0;
 
     const onLayoutHandlerHeader = (event: LayoutChangeEvent): void => {
         const height = event.nativeEvent.layout.height;
-        if (height > 0) {
-            if (height <= maxHeight) {
-                setHeightHeader(height);
-            }
+        if (height > 0 && height <= maxHeight) {
+            setHeightHeader(height);
         }
     };
 
@@ -48,13 +51,12 @@ export const ExpandingDrawer = (props: ExpandingDrawerProps): ReactElement => {
         }
     };
 
+    const containerStyle =
+        props.fullscreenContent && isOpen ? props.styles.containerWhenExpandedFullscreen : props.styles.container;
+
     const renderContent = useCallback((): ReactNode => {
         const content = (
-            <View
-                onLayout={onLayoutHandlerContent}
-                style={!props.fullscreenContent ? props.styles.container : {}}
-                pointerEvents="box-none"
-            >
+            <View onLayout={onLayoutHandlerContent} pointerEvents="box-none">
                 <View
                     onLayout={onLayoutHandlerHeader}
                     style={!isSmallContentValid ? { height: 20 } : {}}
@@ -65,15 +67,10 @@ export const ExpandingDrawer = (props: ExpandingDrawerProps): ReactElement => {
                 {props.largeContent}
             </View>
         );
+
         if (props.fullscreenContent) {
             return (
-                <View
-                    style={[
-                        isOpen ? props.styles.containerWhenExpandedFullscreen : props.styles.container,
-                        { height: fullscreenHeight }
-                    ]}
-                    pointerEvents="box-none"
-                >
+                <View style={[{ height: fullscreenHeight }]} pointerEvents="box-none">
                     {content}
                     {props.fullscreenContent}
                 </View>
@@ -103,25 +100,38 @@ export const ExpandingDrawer = (props: ExpandingDrawerProps): ReactElement => {
             ? [heightContent, heightHeader]
             : [heightHeader];
 
+    const collapsedIndex = 0;
+
+    const onChange = (index: number) => {
+        const hasOpened = lastIndexRef === -1 && index === 0;
+        const hasClosed = index === -1;
+
+        if (hasOpened) {
+            props.onOpen?.();
+            setIsOpen(true);
+        }
+        if (hasClosed) {
+            props.onClose?.();
+            setIsOpen(false);
+        }
+        lastIndexRef = index;
+    };
+
     return (
         <View style={StyleSheet.absoluteFillObject} pointerEvents="box-none">
             {snapPoints.length > 1 && (
                 <BottomSheet
-                    enabledManualSnapping={false}
-                    enabledBottomInitialAnimation
-                    enabledContentTapInteraction={false}
-                    enabledHeaderGestureInteraction={false}
-                    snapPoints={snapPoints}
-                    initialSnap={snapPoints.length - 1}
-                    renderContent={renderContent}
-                    enabledInnerScrolling={false}
-                    onOpenStart={props.onOpen}
-                    onOpenEnd={() => setIsOpen(true)}
-                    onCloseStart={() => {
-                        setIsOpen(false);
-                        props.onClose?.();
-                    }}
-                />
+                    ref={bottomSheetRef}
+                    index={collapsedIndex}
+                    snapPoints={snapPoints.map(p => p + OFFSET_BOTTOM_SHEET)}
+                    onClose={() => setIsOpen(false)}
+                    enablePanDownToClose={false}
+                    onChange={onChange}
+                    animateOnMount
+                    backgroundStyle={containerStyle}
+                >
+                    <BottomSheetView style={{ height: heightHeader }}>{renderContent()}</BottomSheetView>
+                </BottomSheet>
             )}
         </View>
     );